fix(android/theme): MenaColors 用 Color(Int) 替代 Color(ULong) 修 BUG-ANDROID-001
11 处常量从 Color(VALUE: ULong) 改为 Color(VALUE.toInt())，避免
Color(ULong) 重载把低 6 位当作 colorspace ID，导致 Android 13 上
ArrayIndexOutOfBoundsException(length=18, index=55/46/62/...)。

色值数值不变（toInt() 保留 32-bit ARGB），无视觉差异。
保留 *_VALUE: ULong 常量供 JVM 单测复用。

测试：
- testLocalDebugUnitTest --tests "*core.theme.*": BUILD SUCCESSFUL
  (MenaColorsTest 7/7 PASS)
- connectedLocalDebugAndroidTest MenaThemeTest: 8/8 PASS
  (修复前 Round 1: 8/8 AIOOBE; 修复后 Round 2: 全绿)
- 全量 androidTest 由 0/57 → 138/180 PASS，剩余 42 项均为
  非 colorspace 范畴失败（详见 TDS Round 2 章节），无 AIOOBE。

Refs: T-30018 Round 2, QA report tests/report-20260429-081255

+++ app/android/app/src/main/java/com/voice/room/android/core/theme/MenaColors.kt
@@ -25,15 +25,18 @@ object MenaColors {
     const val SYSTEM_MESSAGE_VALUE: ULong          = 0xFFF39C12uL
 
     // ── Compose Color 常量 ────────────────────────
-    val Background: Color            = Color(BACKGROUND_VALUE)
-    val Surface: Color               = Color(SURFACE_VALUE)
-    val SurfaceVariant: Color        = Color(SURFACE_VARIANT_VALUE)
-    val Primary: Color               = Color(PRIMARY_VALUE)
-    val PrimaryBright: Color         = Color(PRIMARY_BRIGHT_VALUE)
-    val OnBackground: Color          = Color(ON_BACKGROUND_VALUE)
-    val OnBackgroundSecondary: Color = Color(ON_BACKGROUND_SECONDARY_VALUE)
-    val OnBackgroundTertiary: Color  = Color(ON_BACKGROUND_TERTIARY_VALUE)
-    val Error: Color                 = Color(ERROR_VALUE)
-    val Success: Color               = Color(SUCCESS_VALUE)
-    val SystemMessage: Color         = Color(SYSTEM_MESSAGE_VALUE)
+    // 注意：使用 .toInt() 强制走 Color(color: Int) 重载，按 sRGB ARGB 解码。
+    // 直接传 ULong 会调到 Color(value: ULong)，把低 6 位当作 colorspace ID，
+    // 在 Android 13 上触发 ArrayIndexOutOfBoundsException(length=18) — BUG-ANDROID-001。
+    val Background: Color            = Color(BACKGROUND_VALUE.toInt())
+    val Surface: Color               = Color(SURFACE_VALUE.toInt())
+    val SurfaceVariant: Color        = Color(SURFACE_VARIANT_VALUE.toInt())
+    val Primary: Color               = Color(PRIMARY_VALUE.toInt())
+    val PrimaryBright: Color         = Color(PRIMARY_BRIGHT_VALUE.toInt())
+    val OnBackground: Color          = Color(ON_BACKGROUND_VALUE.toInt())
+    val OnBackgroundSecondary: Color = Color(ON_BACKGROUND_SECONDARY_VALUE.toInt())
+    val OnBackgroundTertiary: Color  = Color(ON_BACKGROUND_TERTIARY_VALUE.toInt())
+    val Error: Color                 = Color(ERROR_VALUE.toInt())
+    val Success: Color               = Color(SUCCESS_VALUE.toInt())
+    val SystemMessage: Color         = Color(SYSTEM_MESSAGE_VALUE.toInt())
 }
